bug: sorting by date now works on all browsers
Dates were being formatted twice in the data transformer. This was only an issue in safari but either waythere was no need to format the date a second time.

Safari was also adding some margin to the button group, this was fixed by adding a 0 margin. While I was doing this I moved the util folder into lib.